#38 - First REST APIs are available
Other rest APIs work in such a way that you simply always request everything as a list to read the data. This can be later optimized using a pagination logic.

Then we use the objects and manipulate them as needed or use them to create new objects (e.g. a finding) using put request.

If you want the AI to predict the finding you simply send the put request and the AI will then execute its logic and return the adjusted finding entity.